# speed up `tune_tbl()` helper